refactor: extract JiraExportOptions.swift from JiraExportConfig.swift
Splits the Jira lookup-option types (JiraProjectOption,
JiraIssueTypeOption) out of JiraExportConfig.swift so that file holds
only the connection/export configuration types.

Byte-identical move of lines 70-92. Both files import Foundation.

JiraExportConfig.swift: 92 → 68 lines. Tests: 667.
